backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
azure-core==1.38.2
azure-identity==1.25.2
black==26.3.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.3.1
cryptography==46.0.5
fastapi==0.135.1
greenlet==3.3.2
h11==0.16.0
idna==3.11
isort==8.0.1
msal==1.35.1
msal-extensions==1.3.1
mypy_extensions==1.1.0
packaging==26.0
pathspec==1.0.4
platformdirs==4.9.4
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.11.0
pyodbc==5.3.0
python-dotenv==1.2.2
pytokens==0.4.1
requests==2.32.5
SQLAlchemy==2.0.48
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
unicorn==2.1.4
urllib3==2.6.3
uvicorn==0.41.0
